pyproject.toml
[tool.poetry]
name = "chatgptbot"
version = "0.1.0"
description = ""
authors = ["Thomas Prescott <tcprescott@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
discord-py = "^2.2.2"
openai = "^0.27.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
